Reorganize directory structure.
Where two different drivers are available, there is now a file for each
named after the driver type (tlpm, visa, kinesis, apt) which contains
the instrument class.
