+++ qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_K10CR1/__init__.py
@@ -1,0 +1,8 @@
+"""Drivers for Thorlabs K10CR1.
+
+There are two driver versions:
+ - Kinesis
+ - APT
+
+The APT platform has been deprecated by Thorlabs.
+"""
