--- qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_K10CR1/apt/__init__.py
@@ -1,1 +1,0 @@
-from .K10CR1 import K10CR1
