--- qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_K10CR1/kinesis/__init__.py
@@ -1,1 +1,0 @@
-from .Thorlabs_K10CR1 import ThorlabsK10CR1
