+++ qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_MFF10x/__init__.py
@@ -1,0 +1,8 @@
+"""Drivers for Thorlabs MFF10x.
+
+There are two driver versions:
+ - Kinesis
+ - APT
+
+The APT platform has been deprecated by Thorlabs.
+"""
