--- qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_MFF10x/apt/__init__.py
@@ -1,1 +1,0 @@
-from .MFF10x import Thorlabs_MFF10x
