--- qcodes_contrib_drivers/drivers/Thorlabs/Thorlabs_MFF10x/kinesis/__init__.py
@@ -1,1 +1,0 @@
-from .Thorlabs_MFF10x import ThorlabsMFF10x
