--- qcodes_contrib_drivers/drivers/Thorlabs/__init__.py
@@ -1,3 +1,0 @@
-from .Thorlabs_K10CR1.apt import K10CR1
-from .Thorlabs_MFF10x.apt import MFF10x
-from .Thorlabs_PRM1Z8.apt import PRM1Z8
