+++ qcodes_contrib_drivers/drivers/Thorlabs/private/kinesis/__init__.py
@@ -1,0 +1,1 @@
+from .core import KinesisInstrument
